refactor(canvas2d-core): split canvas-rendering-context-2d.ts via install*Methods (1208 → 417 + 6)
Applies the same prototype-augmentation pattern proven by
@gjsify/webgl's webgl-context-base.ts split (PRs #220-#222 era):
each method group is a typed `*Methods` interface declaration-merged
into the base class via `declare module`, plus an
`install*Methods(proto)` function that copies the implementations
onto the prototype.

Layout:

  src/canvas-rendering-context-2d.ts (417 LoC)
    — class shell + constructor + fields + ALL internal helpers
      (_ensureSurface, _apply*, _hasShadow, _deviceToUserDistance,
       _renderShadow) + save/restore + ALL state accessors
       (fillStyle..., shadow*, font/textAlign/textBaseline/direction)
       + _toDataURL + _dispose
    — calls installAllContextMethods(CanvasRenderingContext2D.prototype)
      after class declaration

  src/context/transforms.ts        (144 LoC)
    — translate/rotate/scale/transform/setTransform/getTransform/
      resetTransform
  src/context/path-ops.ts          (151 LoC)
    — beginPath/moveTo/lineTo/closePath/bezier+quadraticCurveTo/
      arc/arcTo/ellipse/rect/roundRect
  src/context/drawing.ts           (319 LoC)
    — fill/stroke/fillRect/strokeRect/clearRect/clip/
      isPointInPath/isPointInStroke/drawImage + getDrawImageSource
  src/context/pixels.ts            (127 LoC)
    — createImageData/getImageData/putImageData
  src/context/text-rendering.ts    (246 LoC)
    — fillText/strokeText/measureText (+ parseFontToDescription /
      createTextLayout / getTextAlignOffset / getTextBaselineOffset
      as module-local helpers — these are NOT exposed on the class)
  src/context/factories.ts         (49 LoC)
    — createLinearGradient/createRadialGradient/createPattern
  src/context/index.ts             (34 LoC)
    — barrel exposing installAllContextMethods(proto)

Important detail: the base file uses a bare side-effect
`import './context/index.js';` AT THE TOP in addition to the named
`import { installAllContextMethods }` at the bottom. tsc strips
value-only named imports from emitted .d.ts (they're only used at
the value level, so the compiler considers them unreferenced for
types), which would drop the `declare module` augmentations from
downstream's view of @gjsify/canvas2d-core's published types. The
side-effect import survives the .d.ts emit, ensuring downstream
packages (@gjsify/canvas2d, @gjsify/dom-elements, framework
consumers) see all method signatures on CanvasRenderingContext2D.

Public surface preserved verbatim — every export from
@gjsify/canvas2d-core works unchanged. Private fields dropped to
no-modifier (`_surface`, `_ctx`, `_state`, etc.) since the split
modules need to access them via `this._state` / `this._ctx`; the
leading-underscore convention still marks them as
implementation-internal. tsc clean across canvas2d-core +
dom-elements; 578 GJS canvas2d-core tests still green.

STATUS.md: new entry under Completed.
